Moving logic for determining pledge, commitment and normal donations into useHook
